chore: erase and rename components
