--- client/src/components/CaseCard/index.js
@@ -1,1 +1,0 @@
-export { default } from './CaseCard'
