--- client/src/components/PostCard/index.js
@@ -1,1 +1,0 @@
-export { default } from './PostCard'
