+++ client/src/pages/case/[slug].js
@@ -6,7 +6,7 @@ import { groq } from 'next-sanity'
 import client from '../../sanity-client'
 import BlockContent from '../../components/BlockContent'
 import MetaTags from '../../components/MetaTags/MetaTags'
-import CaseList from '../../sections/case/CaseList1'
+import CaseList from '../../sections/case/CaseList'
 import CTA from '../../sections/case/CTA'
 import { usePreviewSubscription } from '../../lib/sanity'
 import { getClient } from '../../lib/sanity.server'
